Move logic for opening log files to `brioche::jobs` module
