+++ crates/brioche/src/jobs/logs.rs
@@ -1,14 +1,12 @@
 use std::path::PathBuf;
 
-use brioche_core::{
-    process_events::{
-        PROCESS_EVENT_MAGIC,
-        display::{DisplayEventsOptions, display_events},
-    },
-    utils::io::NotSeekable,
-};
+use brioche_core::process_events::display::{DisplayEventsOptions, display_events};
 use clap::Parser;
 use notify::Watcher as _;
+
+use crate::jobs::log_file_reader_from_stdin;
+
+use super::log_file_reader_from_path;
 
 #[derive(Debug, Parser)]
 pub struct LogsArgs {
@@ -28,48 +26,16 @@ pub struct LogsArgs {
     follow: bool,
 }
 
-const ZSTD_FRAME_MAGIC: &[u8] = &[0x28, 0xB5, 0x2F, 0xFD];
-const ZSTD_SKIPPABLE_FRAME_MAGIC: &[u8] = &[0x2A, 0x4D, 0x18];
-
-trait ReadSeek: std::io::Read + std::io::Seek {}
-impl<T: std::io::Read + std::io::Seek> ReadSeek for T {}
-
 pub fn logs(args: &LogsArgs) -> anyhow::Result<()> {
-    let input: Box<dyn ReadSeek> = if args.path.to_str() == Some("-") {
+    let input = if args.path.to_str() == Some("-") {
         anyhow::ensure!(
             !args.follow,
             "cannot specify --follow when reading from stdin"
         );
 
-        let mut stdin = std::io::stdin().lock();
-        let format = detect_format(&mut stdin)?;
-
-        match format {
-            LogFileFormat::Zstd => {
-                let decoder = zstd_framed::ZstdReader::builder_buffered(stdin).build()?;
-                Box::new(NotSeekable(decoder))
-            }
-            LogFileFormat::Bin => Box::new(NotSeekable(stdin)),
-        }
+        log_file_reader_from_stdin()?
     } else {
-        let file = std::fs::File::open(&args.path)?;
-        let mut buf_reader = std::io::BufReader::new(file);
-        let format = detect_format(&mut buf_reader)?;
-
-        match format {
-            LogFileFormat::Zstd => {
-                let seek_table = zstd_framed::table::read_seek_table(&mut buf_reader)
-                    .ok()
-                    .flatten();
-                let mut decoder = zstd_framed::ZstdReader::builder(buf_reader);
-                if let Some(seek_table) = seek_table {
-                    decoder = decoder.with_seek_table(seek_table);
-                }
-                let decoder = decoder.build()?;
-                Box::new(decoder)
-            }
-            LogFileFormat::Bin => Box::new(buf_reader),
-        }
+        log_file_reader_from_path(&args.path)?
     };
 
     let mut reader = brioche_core::process_events::reader::ProcessEventReader::new(input)?;
@@ -102,36 +68,4 @@ pub fn logs(args: &LogsArgs) -> anyhow::Result<()> {
     )?;
 
     Ok(())
-}
-
-fn detect_format(reader: &mut impl std::io::BufRead) -> anyhow::Result<LogFileFormat> {
-    let buf = reader.fill_buf()?;
-
-    let process_event_magic = PROCESS_EVENT_MAGIC.as_bytes();
-    let process_event_magic_partial_len = process_event_magic.len().min(buf.len());
-    let process_event_magic_partial = &process_event_magic[0..process_event_magic_partial_len];
-    if buf.starts_with(process_event_magic_partial) {
-        return Ok(LogFileFormat::Bin);
-    }
-
-    let zstd_frame_magic_partial_len = ZSTD_FRAME_MAGIC.len().min(buf.len());
-    let zstd_frame_magic_partial = &ZSTD_FRAME_MAGIC[0..zstd_frame_magic_partial_len];
-    if buf.starts_with(zstd_frame_magic_partial) {
-        return Ok(LogFileFormat::Zstd);
-    }
-
-    let zstd_skippable_frame_magic_partial_len = ZSTD_SKIPPABLE_FRAME_MAGIC.len().min(buf.len());
-    let zstd_skippable_frame_magic_partial =
-        &ZSTD_SKIPPABLE_FRAME_MAGIC[0..zstd_skippable_frame_magic_partial_len];
-    if buf.starts_with(zstd_skippable_frame_magic_partial) {
-        return Ok(LogFileFormat::Zstd);
-    }
-
-    anyhow::bail!("could not detect log file format");
-}
-
-#[derive(Debug, Clone, Copy)]
-enum LogFileFormat {
-    Zstd,
-    Bin,
 }
